refactor: RecentView 컴포넌트로 나누기
